# Conflicts:
#	client-app/src/container/intake/list/IntakesListToolbar.js
#	client-app/src/container/project/details/revenueKwh/details/revenue-distribution/RevenuesKwhDistributionForm.js
#	config/app.php
#	customer-portal-app/build/asset-manifest.json
#	customer-portal-app/build/index.html
#	customer-portal-app/build/service-worker.js
#	customer-portal-app/build/static/css/main.8bd1e48e.chunk.css
#	customer-portal-app/build/static/js/2.aff08069.chunk.js
#	customer-portal-app/build/static/js/2.aff08069.chunk.js.LICENSE.txt
#	customer-portal-app/build/static/js/2.aff08069.chunk.js.map
#	public/js/100.3b7d9a7df255f2f9f330.js
#	public/js/101.9ad2647f5e7c91ae2aa6.js
#	public/js/105.b807e2353287120a0691.js
#	public/js/106.db8e041136436fe47860.js
#	public/js/108.51229e97d379cca412ef.js
#	public/js/110.83d6d05bc9a85e044869.js
#	public/js/112.3723ea22f5606f90c6f2.js
#	public/js/113.5fd6d8b751248f0fe508.js
#	public/js/114.b7dc547091cc710c30fe.js
#	public/js/115.ea3832b9d75f8fc37753.js
#	public/js/116.6f8dac8bf05e8179a7c5.js
#	public/js/117.c8df225e82fbdbc71d3f.js
#	public/js/118.ed6764283e00ad0c5cfe.js
#	public/js/119.5a33dd847c15fa733aa8.js
#	public/js/12.1cce67c16c9fadfdf6b1.js
#	public/js/125.979c92d551fb9beb3363.js
#	public/js/126.516052253eabc80c210b.js
#	public/js/129.23762065fa6f9f045830.js
#	public/js/13.e78e90a133112cea4ba5.js
#	public/js/130.2743564bd83c863e59bc.js
#	public/js/133.4b3473a02acad712ecd1.js
#	public/js/135.e059197c34769e0ad130.js
#	public/js/138.22042c2b44721c457dc0.js
#	public/js/14.b2f988b8828326b2b896.js
#	public/js/140.e0f817038a8a6c5afd12.js
#	public/js/141.0a7d07b1ac9a377292e4.js
#	public/js/142.a4b9b01d5b1b7bd18b56.js
#	public/js/144.5ae0d44240e22fa19b11.js
#	public/js/145.9067447796c7d3ca9d99.js
#	public/js/146.7017e594e3e41cba0098.js
#	public/js/147.7a888263d3abd14bbbeb.js
#	public/js/148.a9b651b3d9f885da3675.js
#	public/js/149.6b88b1c8133a3f1ca028.js
#	public/js/151.c4417cfa95fbcb931452.js
#	public/js/153.c58faaa3c2441f58da92.js
#	public/js/154.efe7a93de9ee81a2946c.js
#	public/js/156.f597d4729b1cbd180923.js
#	public/js/17.eb8cd7a9e3fc2fd4fc6e.js
#	public/js/18.50415b914d5e95950c7f.js
#	public/js/19.7a0161e006ef82925684.js
#	public/js/20.9380fc8cf44ff5bed7a8.js
#	public/js/21.178dc870792b22581d8c.js
#	public/js/22.8d35bf2051bf0c4549c7.js
#	public/js/23.db7491efbc4193674a65.js
#	public/js/24.91fda5be0afde932401f.js
#	public/js/25.d97b3eb0f85aa44f282c.js
#	public/js/26.4b78c7867c980fd6f222.js
#	public/js/27.b6e2c808020a3af8338d.js
#	public/js/28.931e422820e3014de5b9.js
#	public/js/29.abbaff51ca7ec92c22a3.js
#	public/js/30.ac7a92f85f4874a9e640.js
#	public/js/31.21932ed5a28d6d1ca8c7.js
#	public/js/33.2577d4ef3c7748710d5e.js
#	public/js/39.ea34dd5af265b0946c8b.js
#	public/js/43.1cc12a3dddf65e0ac921.js
#	public/js/44.b4261fce5db119f9c26d.js
#	public/js/47.53fc881bdcfcfa29c948.js
#	public/js/48.48326a96e61e6316a19d.js
#	public/js/50.953850386c76f425da9e.js
#	public/js/51.63f02877b0083acc5d38.js
#	public/js/52.7aeacbadb376e384c3ba.js
#	public/js/53.1a527e6fbecc96e941f8.js
#	public/js/56.fdbabd702781b1ef9eb0.js
#	public/js/59.bb324ef1374a6c7f1b0b.js
#	public/js/6.924fb2de7d675c0d8dcb.js
#	public/js/60.ddbbfbbeafa885aa1e34.js
#	public/js/62.9276821e99a38ef5329d.js
#	public/js/63.715e7fb6f150945bb39e.js
#	public/js/64.0b736c1ac7b3bd7dd7bc.js
#	public/js/66.1578f1b500f6693a1e9a.js
#	public/js/67.5786f4c3a4ea8076074e.js
#	public/js/7.cd8ecb85aa65b532c41f.js
#	public/js/72.5bd028b989378c2fb956.js
#	public/js/73.d1760fd5b06ed7b359a1.js
#	public/js/74.51196f37414d1c405329.js
#	public/js/75.398b840f36c4149307fc.js
#	public/js/76.5cc1e9de7f24dea11e6d.js
#	public/js/77.928bf43efa459002cea1.js
#	public/js/78.30530332e3eaf56a6119.js
#	public/js/79.c30b843f280657c93b97.js
#	public/js/8.720b67d6f50af0003a6b.js
#	public/js/80.6a0659e9398bd0075bb7.js
#	public/js/82.72d06a6ca0169a7b4794.js
#	public/js/84.85341e4265ba77e04482.js
#	public/js/85.a35dd87e47cca8b787d1.js
#	public/js/86.3fdfc721ff4fe8f7b503.js
#	public/js/87.4b93699ff14ad8adbcc5.js
#	public/js/88.fb40b0b87784969e756e.js
#	public/js/89.ca4c0eedca3be03919b5.js
#	public/js/91.338e338ead3f2c592e65.js
#	public/js/92.bd0db333346295f05023.js
#	public/js/93.a62a32a57b54bcc55d89.js
#	public/js/94.3dc6a002ceebf99bc926.js
#	public/js/97.d4584478dfb414f48c0a.js
#	public/js/98.9232acf105a4f8601429.js
#	public/js/vendors~bundle.5a2c18a77b3b756f72b8.js
#	resources/views/welcome.blade.php

+++ app/Http/Controllers/Api/Project/RevenuesKwhController.php
@@ -178,8 +178,13 @@ class RevenuesKwhController extends ApiController
             ->orderBy('date_begin')->get());
     }
 
+    public function recalculateRevenuesDistribution(RevenuesKwh $revenuesKwh)
+    {
+        $this->saveParticipantsOfDistribution($revenuesKwh);
+    }
+
     //todo WM: dit naar job verplaatsen ?!
-    public function saveParticipantsOfDistribution(RevenuesKwh $revenuesKwh)
+    protected function saveParticipantsOfDistribution(RevenuesKwh $revenuesKwh)
     {
         set_time_limit(300);
 
@@ -241,8 +246,11 @@ class RevenuesKwhController extends ApiController
                 $distributionKwh->participations_quantity = $quantityOfParticipations;
                 $distributionKwh->save();
             } else {
-                $deleteRevenueDistributionKwh = new DeleteRevenueDistributionKwh($distributionKwh);
-                $deleteRevenueDistributionKwh->delete();
+                $revenuePartsKwhHasConfirmed = $revenuesKwh->partsKwh()->where('confirmed', true)->exists();
+                if(!$revenuePartsKwhHasConfirmed){
+                    $deleteRevenueDistributionKwh = new DeleteRevenueDistributionKwh($distributionKwh);
+                    $deleteRevenueDistributionKwh->delete();
+                }
             }
         }
     }
